Fix currentMarker initialization
What was thought to be an event delegation bug turned out to be an issue
in how I was initializing the variable in the Game module.
Initializing the variable and adding it to the global namespace
removed the necessity for the initialization function within the
Game module. Also added checks for the negative instead of the positive
in the event handler function in the displayController module to allow
for more shallow code.